Formatted footer and end sign up
Formatted the footer to align in center, formatted section below the quote and it is perfect. No changes needed to either. Quote needs to be centered on desktop fullscreen as well as the quote source needs to be better aligned. Need to put the header bar and hero into a div together to align properly.